Create 5oZ16Y.cpp
solved issue #504


To calculate the volume of a dodecahedron, you can use the following formula:
Volume = (15 + 7 * sqrt(5)) / 4 * a^3

In this code, we define a function calculateDodecahedronVolume that takes the length of the edge of the dodecahedron as an input and returns its volume. In the main function, we take the input from the user, validate that the edge length is non-negative, and then calculate and display the volume.We use a `std::unique_ptr` to dynamically allocate memory for the edge length of the dodecahedron.

Signed-off-by: Subhadip956425 <122253446+Subhadip956425@users.noreply.github.com>