Create eeOLZZ.py
solved issue #496 

In this code, we define a function called `calculate_cylinder_volume` that takes two parameters: `radius` and `height`. The function uses the formula `V = πr²h` to calculate the volume of the cylinder, where `V` is the volume, `r` is the radius, and `h` is the height. The function returns the calculated volume.
We need to import the `math` module to use the `pi` constant and the `sqrt` function. If you don't import the `math` module, you will get a `NameError`.

Signed-off-by: Subhadip956425 <122253446+Subhadip956425@users.noreply.github.com>